typf-tester/requirements.txt
# Typf Tester Python Dependencies
# Community project by FontLab https://www.fontlab.org/

# CLI framework
fire>=0.6.0

# Optional: Image verification (PNG)
Pillow>=10.0.0
